+++ code/compiler/builder.node/Builder.mts
@@ -2,7 +2,7 @@ import { build } from './Builder.build.mjs';
 import { clean } from './Builder.clean.mjs';
 import { test } from './Builder.test.mjs';
 import { FindUtil as Find } from './common/index.mjs';
-import { Dependencies } from './op/Dependencies.mjs';
+import { Dependencies } from './operations/Dependencies.mjs';
 import { Paths } from './Paths.mjs';
 
 /**
