+++ code/compiler/builder.node/operations/Manifest.mts
