+++ code/compiler/builder.node/operations/Vite.mts
